% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity.R
\name{annuity}
\alias{annuity}
\title{Annuitaetenbarwert}
\usage{
annuity(annuity = 1, rate = 0.02, periods = 1)
}
\arguments{
\item{annuity}{Hoehe der Annuitaet, standardmaessig 1}

\item{rate}{Zinssatz in Dezimalform, standardmaessig 0.02}

\item{periods}{Laufzeit in Jahren, standardmaessig 1}
}
\value{
Barwert der Annuitaten bzw. Annuitaetenbarwertfaktor
}
\description{
Ermittelt den Barwert einer Annuitaet basierend auf Laufzeit und Zinssatz.
Wird keine Annuitaet angegeben, wird diese mit 1 angenommen und folglich der
Annuitaetenbarwertfaktor ermittelt.
}
\examples{
annuity(annuity = 50, rate = 0.02, periods = 20) # 817.5717
annuity(rate = 0.02, periods = 20) # 16.35143
}
