% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv.R
\name{fv}
\alias{fv}
\title{Endwert}
\usage{
fv(payment = 1, rate = 0.02, periods = 1)
}
\arguments{
\item{payment}{Heutige Zahlung, standardmaessig 1}

\item{rate}{Zinssatz in Dezimalform, standardmaessig 0.02}

\item{periods}{Laufzeit in Jahren, standardmaessig 1}
}
\value{
Endwert bzw. Aufzinsungsfaktor
}
\description{
Ermittelt den Endwert einer heutigen Zahlung basierend auf Laufzeit und
Zinssatz. Wird keine Zahlung angegeben, wird diese mit 1 angenommen und
folglich der Aufzinsungsfaktor ermittelt.
}
\examples{
fv(payment = 50, rate = 0.02, periods = 20) # 74.29737
fv(rate = 0.02, periods = 20) # 1.485947
}
