% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price.R
\name{price}
\alias{price}
\title{Preis einer Anleihe}
\usage{
price(cash_flows, ytm)
}
\arguments{
\item{cash_flows}{Vektor der Cashflows}

\item{ytm}{Yield-to-Maturity in Dezimalform}
}
\value{
Preis der Anleihe
}
\description{
Berechnet den Preis einer Anleihe auf Basis der uebergebenen Cashflows und
Yield-to-Maturity.
}
\examples{
price(c(-95, 2, 2, 2, 2, 102), 0.01)
}
