% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.R
\name{pv}
\alias{pv}
\title{Barwert}
\usage{
pv(payment = 1, rate = 0.02, periods = 1)
}
\arguments{
\item{payment}{Zukuenftige Zahlung, standardmaessig 1}

\item{rate}{Zinssatz in Dezimalform, standardmaessig 0.02}

\item{periods}{Laufzeit in Jahren, standardmaessig 1}
}
\value{
Barwert bzw. Barwertfaktor
}
\description{
Ermittelt den Barwert einer zukuenftigen Zahlung basierend auf Laufzeit und
Zinssatz. Wird keine Zahlung angegeben, wird diese mit 1 angenommen und
folglich der Barwertfaktor ermittelt.
}
\examples{
pv(payment = 500, rate = 0.03, periods = 5) # 431.3044
pv(rate = 0.03, periods = 5) # 0.8626088
}
