% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytm.R
\name{ytm}
\alias{ytm}
\title{Yield-to-Maturity}
\usage{
ytm(cash_flows)
}
\arguments{
\item{cash_flows}{Vektor der Cashflows}
}
\value{
Yield-to-Maturity in Dezimalform
}
\description{
Berechnet die Yield-to-Maturity (Endfaelligkeitsrendite) einer Anleihe auf
Basis der uebergebenen Cashflows. Die Approximation erfolgt mittels
stats::uniroot() im Intervall -1:1, wodurch auch negative Renditen zulaessig
sind.
}
\examples{
ytm(c(-95, 2, 2, 2, 2, 102))
}
